Moved consent doc into HTML file
